libc standards reset and C89 math core
* Normalize libc headers to standards-first APIs and types; move syscall numbers into new <apheleia/syscall.h> and update kernel/userspace dispatch points accordingly
* Add missing baseline C headers and runtime pieces (locale C/POSIX support, setjmp, assert runtime, wide/char type stubs, stdio and dirent surface cleanup)
* Fix libc behavioral mismatches in ctype/string/qsort/printf+scanf/time paths and migrate libc_usr wrappers plus user tools to the new interfaces
* Replace arch-specific libc math plumbing with a single self-contained C89 math implementation and keep complex support explicitly disabled
* Clean up related build and userland integration paths (syscall numbering, sleep/nanosleep routing, account/proc helpers, and command-side API updates)

+++ libs/libc/sys/types.h
@@ -5,12 +5,12 @@
 
 typedef ptrdiff_t ssize_t;
 
-typedef ssize_t pid_t;
-typedef ssize_t tid_t;
-typedef ssize_t uid_t;
-typedef ssize_t gid_t;
+typedef int pid_t;
+typedef int tid_t;
+typedef unsigned int uid_t;
+typedef unsigned int gid_t;
 
-typedef ssize_t off_t;
+typedef long off_t;
 
 typedef uint32_t mode_t;
